% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diageo_pe.R
\docType{data}
\name{diageo_pe}
\alias{diageo_pe}
\title{Data for daily price earning ratio of Diageo PLC}
\format{
A tibble with 2562 rows and 2 variables:
\describe{ The data is download from Capital IQ and the ratio is calculated as the Price as a proportion of the Diluted Earnings per share
\item{Date}{ddtm daily dates}
\item{PE}{dbl Price earnings ratio for Diageo}
}
}
\source{
\url{https://raptor.qub.ac.uk/genericInstruction.php?&suborg=qub&resourceId=41&url=http://www.capitaliq.com/}
}
\usage{
diageo_pe
}
\description{
Data for daily price earning ratio of Diageo PLC
}
\keyword{datasets}
