% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glen_m_r.R
\docType{data}
\name{glen_m_r}
\alias{glen_m_r}
\title{Glencore PLC monthly log returns}
\format{
a monthly ts object:
\describe{log returns from MAY 2011 to DEC 2020}
}
\source{
\url{Yahoo Finance}
}
\usage{
glen_m_r
}
\description{
Glencore PLC monthly log returns
}
\keyword{datasets}
