% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\docType{data}
\name{indices}
\alias{indices}
\title{Data for daily stock prices indices and major currency exchange rates}
\format{
A tibble with 15651 rows and 19 variables:
\describe{ The data is download from Refinitiv Datastream
\item{}{dttm daily date spannig back to 1960 for some of the indices}
\item{"CHINA A-DS MARKET $ - PRICE INDEX"}{dbl DataStreams China A Share Price Index}
\item{"DAX 30 PERFORMANCE - PRICE INDEX" }{dbl Germany DAX 30 Price Index}
\item{"DOW JONES INDUSTRIALS - PRICE INDEX"}{dbl DJI Price Index}
\item{"FTSE ALL SHARE - PRICE INDEX"}{FTSE all share price index}
\item{"GBP TO EUR BOE - EXCHANGE RATE"}{Bank of England Pound Euro Exchange Rate nicknamed by traders the Chunnel}
\item{"GBP TO USD BOE - EXCHANGE RATE"}{Bank of England Pound US Dollar Exchange Rate nicknamed by traders  Cable}
\item{"GERMANY-DS NON-FIN.EX-RESOR - PRICE INDEX"}{Datastream Germany stock market price index excluding financials}
\item{"IRELAND-DS MARKET $ - PRICE INDEX"}{Datastream Ireland stock market price index}
\item{"JAPAN-DS NON-FIN.EX-RESOR - PRICE INDEX"}{Datastream Japan stock market price index excluding financials}
\item{"NIKKEI 225 STOCK AVERAGE - PRICE INDEX"}{Nikkei 225 Japanese stock market price index}
\item{"RUSSELL 2000 - PRICE INDEX"}{The Russell 2000 Index measures the performance of the small-cap segment of the US equity universe.
The Russell 2000 Index is a subset of the Russell 3000 Index representing approximately 10\% of the total market capitalization of that index.
It includes approximately 2,000 of the smallest securities based on a combination of their market cap and current index membership.}
\item{"S&P 500 COMPOSITE - PRICE INDEX"}{S&P 500 price index}
\item{"S&P 500 COMPOSITE DS CALCULATED - PRICE INDEX"}{Datstream calculate S&P 500 price index}
\item{"UK £ BROAD INDEX JAN 05=100 BOE - TRADE WEIGHTED"}{The sterling BROAD exchange rate index is a measure of the overall change in the trade-weighted exchange value of sterling, calculated by weighting together bilateral exchange rates. It is designed to measure changes in the price competitiveness of traded goods and services, and so the weights reflect trade flows in manufactured goods and services.}
\item{"UK £ EFFECTIVE INDEX JAN 05=100 BOE - TRADE WEIGHTED"}{The sterling Effective exchange rate index is a measure of the overall change in the trade-weighted exchange value of sterling, calculated by weighting together bilateral exchange rates. It is designed to measure changes in the price competitiveness of traded goods and services, and so the weights reflect trade flows in manufactured goods and services.}
\item{"UK-DS NON-FIN.EX-RESOR - PRICE INDEX"}{Datastream UK stock market price index excluding financials}
\item{"USD TO GBP (BOE) - EXCHANGE RATE"}{Bank of England US Dollar Pound Exchange Rate nicknamed by traders  Cable}
\item{"WORLD-DS Market - PRICE INDEX"}{Datastream world stock market price index}
}
}
\source{
\url{https://raptor.qub.ac.uk/genericInstruction.php?&suborg=qub&resourceId=41&url=http://www.capitaliq.com/, https://research.ftserussell.com/Analytics/FactSheets/temp/c191f598-37d9-4dad-94d7-64545f3317ff.pdf}
}
\usage{
indices
}
\description{
Data for daily stock prices indices and major currency exchange rates
}
\keyword{datasets}
