% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monte_carlo_paths.R
\name{monte_carlo_paths}
\alias{monte_carlo_paths}
\title{Monte Carlo Simulation of Stock Price Paths}
\usage{
monte_carlo_paths(
  days = 250,
  mean = 1.001,
  sd = 0.005,
  S0 = 40,
  number_of_possible_futures = 1000
)
}
\arguments{
\item{days}{Number of days to simulate.}

\item{mean}{Mean of the daily returns.}

\item{sd}{Standard deviation of the daily returns.}

\item{S0}{Initial stock price.}

\item{number_of_possible_futures}{Number of paths to simulate.}
}
\value{
A tibble containing the simulated paths.
}
\description{
Simulates multiple possible future paths of a stock price using a Monte Carlo method.
}
\examples{
monte_carlo_paths(days = 250, mean = 1.001, sd = 0.005, S0 = 40, number_of_possible_futures = 1000)
}
