% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale.R
\name{rescale_variable}
\alias{rescale_variable}
\title{Rescale a Numeric Vector}
\usage{
rescale_variable(x, a, b)
}
\arguments{
\item{x}{A numeric vector.}

\item{a}{The lower bound of the new scale.}

\item{b}{The upper bound of the new scale.}
}
\value{
A numeric vector rescaled to be within \link{a, b}.
}
\description{
Rescales a given numeric vector to be bounded between two specified values.
}
\examples{
data <- c(1, 2, 3, 4, 5)
rescale_variable(data, 0, 10)
}
