% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsfe-package.R
\docType{package}
\name{tsfe-package}
\alias{tsfe}
\alias{tsfe-package}
\title{tsfe: What the Package Does (One Line, Title Case)}
\description{
used to teach time series financial econometrics.
}
\author{
\strong{Maintainer}: Barry Quinn \email{bquinn@qub.ac.uk} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
